pyproject.toml
[project]
name = "the-oracle"
version = "0.1.0"
description = "A recursive deep-research agent system"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "tavily-python>=0.5.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "supabase>=2.0.0",
    "postgrest>=0.16.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
